requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
mysql-connector-python
python-dotenv
